+++ src/Http/Controllers/V1/Profile/WebController.php
@@ -5,6 +5,7 @@ namespace Werify\Account\Laravel\Http\Controllers\V1\Profile;
 use Illuminate\Routing\Controller;
 use Werify\Account\Laravel\Http\Requests\V1\Profile\UpdateRequest;
 use Werify\Account\Laravel\Jobs\V1\Profile\UpdateJob;
+use Werify\Account\Laravel\WAccount;
 
 class WebController extends Controller
 {
@@ -32,7 +33,7 @@ class WebController extends Controller
                 }
             }
 
-            return redirect($this->replaceLanguageInUrl(url()->previous(), $r->input('language')));
+            return redirect(WAccount::replaceLanguageInUrl(url()->previous(), $r->input('language')));
         } catch (\Exception $e) {
             if (config('waccount.debug')) {
                 throw $e;
@@ -62,46 +63,5 @@ class WebController extends Controller
         }
     }
 
-    public function replaceLanguageInUrl($url, $newLanguage = null)
-    {
-        if (empty($newLanguage)) {
-            return $url;
-        }
-        $parsedUrl = parse_url($url);
 
-        // Replace language code in path
-        if (isset($parsedUrl['path'])) {
-            $pathSegments = explode('/', trim($parsedUrl['path'], '/'));
-            if (count($pathSegments) > 0) {
-                $pathSegments[0] = $newLanguage;
-                $parsedUrl['path'] = '/'.implode('/', $pathSegments);
-            }
-        }
-
-        // Replace language code in query parameters
-        if (isset($parsedUrl['query'])) {
-            parse_str($parsedUrl['query'], $queryParams);
-            if (count($queryParams) > 0) {
-                $queryParams['lang'] = $newLanguage;
-                $parsedUrl['query'] = http_build_query($queryParams);
-            }
-        }
-
-        // Reconstruct the modified URL
-        $modifiedUrl = $parsedUrl['scheme'].'://'.$parsedUrl['host'];
-        if (isset($parsedUrl['port'])) {
-            $modifiedUrl .= ':'.$parsedUrl['port'];
-        }
-        if (isset($parsedUrl['path'])) {
-            $modifiedUrl .= $parsedUrl['path'];
-        }
-        if (isset($parsedUrl['query'])) {
-            $modifiedUrl .= '?'.$parsedUrl['query'];
-        }
-        if (isset($parsedUrl['fragment'])) {
-            $modifiedUrl .= '#'.$parsedUrl['fragment'];
-        }
-
-        return $modifiedUrl;
-    }
 }