+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2023-08-23
+
+### Added
+ - Add `WAccount` class for general methods and helpers.
+ - Add `initializeIndex` method in `WAccount` class for initialize index page with support of language based on user location.
+
 ## [Unreleased] - 2023-08-22
 
 ### Added